+++ CHANGES.rst
@@ -135,6 +135,10 @@ Release History
   `#862 <https://github.com/nengo/nengo/pull/862>`_)
 - Added SPA wrapper for circular convolution networks, ``spa.Bind``
   (`#849 <https://github.com/nengo/nengo/pull/849>`_)
+- Added the ``Voja`` (Vector Oja) learning rule type, which updates an
+  ensemble's encoders to fire selectively for its inputs. (see
+  ``examples/learning/learn_associations.ipynb``).
+  (`#727 <https://github.com/nengo/nengo/issues/727>`_)
 
 **Bug fixes**